+++ acceptance-tests/test-plugins/src/main/java/org/hyperledger/besu/tests/acceptance/plugins/TestRpcEndpointServicePlugin.java
@@ -12,7 +12,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.hyperledger.besu.plugins;
+package org.hyperledger.besu.tests.acceptance.plugins;
 
 import static com.google.common.base.Preconditions.checkArgument;
